added floating point numbers to item prices

+++ finalProject/Shared/Models/CartManager.swift
@@ -9,7 +9,8 @@ import Foundation
 
 class CartManager: ObservableObject {
     @Published private(set) var products: [Product] = []
-    @Published private(set) var total: Int = 0
+    @Published private(set) var total: Double = 0.0
+    
     
     func addToCart(product: Product) {
         products.append(product)
